refactor(input): 컨트롤러의 입력 기능을 view 계층으로 추출
